Use std::sqrt instead of sqrt
* In response to request by Brian Ripley who notes:

Subject: ANNperf.h in packages RANN, FNN, msr, nnDiag and yaImpute

You all seem to have copied the same broken piece of C++.  sqrt is in the std:: namespace, so you need to add

using std::sqrt;

to that file.

Can we please have corrected updates (and please report this bug to the authors of ANN).